+++ src/test/java/chess/web/service/ChessServiceTest.java
@@ -8,6 +8,8 @@ import chess.web.controller.dto.MoveDto;
 import chess.web.controller.dto.ScoreDto;
 import chess.web.dao.PieceDao;
 import chess.web.dao.RoomDao;
+import chess.web.service.mock.MockPieceDao;
+import chess.web.service.mock.MockRoomDao;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.transaction.annotation.Transactional;
