+++ src/test/java/chess/web/service/mock/MockRoomDao.java
@@ -1,4 +1,4 @@
-package chess.web.service;
+package chess.web.service.mock;
 
 import chess.domain.board.Team;
 import chess.domain.board.Turn;
